+++ chatroom/global.hpp
@@ -50,6 +50,40 @@ enum class MsgDirection : std::uint8_t
   Received = 1,  // 收到的消息
 };
 
+// 联系人分组
+enum class ContactGroup : std::uint8_t
+{
+  NewFriends = 0,  // 新的朋友
+  A,               // 字母组
+  B,
+  C,
+  D,
+  E,
+  F,
+  G,
+  H,
+  I,
+  J,
+  K,
+  L,
+  M,
+  N,
+  O,
+  P,
+  Q,
+  R,
+  S,
+  T,
+  U,
+  V,
+  W,
+  X,
+  Y,
+  Z,
+  Hash,  // # 组
+  Count  // 分组总数
+};
+
 namespace UIConstants
 {
 // mainwindow.hpp
@@ -59,14 +93,28 @@ constexpr int WindowChatWidth = 1080;   // 聊天窗口宽度
 constexpr int WindowChatHeight = 720;   // 聊天窗口高度
 
 // chatdialog.hpp
-constexpr int ChatUserItemHeight = 65;    // 联系人列表项高度
-constexpr int SearchUserItemHeight = 65;  // 搜索用户列表项高度
-constexpr int LoadingItemHeight = 65;     // 加载中列表项高度
+constexpr int ChatUserItemHeight = 65;      // 联系人列表项高度
+constexpr int SearchUserItemHeight = 65;    // 搜索用户列表项高度
+constexpr int LoadingItemHeight = 65;       // 加载中列表项高度
+constexpr int ContactUserItemHeight = 65;   // 联系人列表项高度
+constexpr int ContactGroupItemHeight = 25;  // 联系人分组项高度
+constexpr int FriendApplyItemHeight = 70;   // 好友申请列表项高度
 
 // chatmsgitem.hpp
 constexpr int ChatMsgItemAvatarSize = 50;    // 头像大小
 constexpr int ChatMsgItemTriangleWidth = 8;  // 小三角宽度
 constexpr int ChatMsgItemBubbleRadius = 6;   // 气泡圆角半径
 }  // namespace UIConstants
+
+// 拼音相关的操作
+namespace PinyinUtils
+{
+constexpr ushort kChineseStart = 0x4E00;
+constexpr ushort kChineseEnd = 0x9FA5;
+
+extern const std::string_view kPinyinDict;
+
+char GetFirstPinyinLetter(QChar cha);
+}  // namespace PinyinUtils
 
 #endif  // GLOBAL_HPP