+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2026-02-02]
+
+### ✨ Features
+- Add contact grouping and friend application UI
+
 ## [2026-01-10]
 
 ### ✨ Features